backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
pydantic==2.11.3
pydantic-settings==2.9.1
aiosqlite==0.21.0
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
bcrypt>=4.0.0,<5.0.0
google-genai==1.10.0
chromadb==0.6.3
pymupdf==1.25.5
pytesseract==0.3.13
resend==2.10.0
pandas==2.2.3
python-multipart==0.0.20
